+++ src/extract/languages/javascript.rs
@@ -63,6 +63,7 @@ pub static JS_CONFIG: LanguageConfig = LanguageConfig {
     call_function_field: "function",
     call_accessor_node_types: &["member_expression"],
     call_accessor_field: "property",
+    call_object_field: None,
     function_boundary_types: &[
         "function_declaration",
         "arrow_function",